Notify web3inbox on x-chain & multilink claiming
