+++ src/components/claim/views/success.view.tsx
@@ -35,25 +35,14 @@ export function ClaimSuccessView({ txHash, claimDetails, senderAddress }: _const
     useEffect(() => {
         router.prefetch('/send')
         gaEventTracker('peanut-claimed', 'success')
-        sendNotification(claimDetails[0])
+        sendNotification()
     }, [])
 
-    const sendNotification = async (linkDetails: interfaces.ILinkDetails) => {
-        console.log('sendNotification', senderAddress)
-        const accounts = [`eip155:1:${senderAddress}` ?? '']
+    const sendNotification = async () => {
         const chainName = chainDetails.find(
-            (detail) => detail.chainId.toString() === linkDetails.chainId.toString()
+            (detail) => detail.chainId === claimDetails[0].chainId
         )?.name
-        const notification = {
-            title: 'Peanut Protocol',
-            body: `Your link has been claimed on ${chainName} by ${address ?? ''}`,
-            url: undefined,
-            type: '2aee6e5f-091d-444e-96cd-868ba2ddd0e7',
-        }
-        utils.sendNotification({
-            notification,
-            accounts,
-        })
+        utils.sendNotification(senderAddress, address, chainName)
     }
 
     return (
