+++ src/utils/notifications.utils.ts
@@ -5,8 +5,22 @@ interface INotification {
     type: string
 }
 
-export const sendNotification = async (notificationPayload: { accounts: string[]; notification: INotification }) => {
+export const sendNotification = async (senderAddress: string, recipientAddress: string | undefined, chainName: string) => {
+    console.log('sendNotification', senderAddress)
+    const accounts = [`eip155:1:${senderAddress}` ?? '']
+    const notification: INotification = {
+        title: 'Peanut Protocol',
+        body: `Your link has been claimed on ${chainName} by ${recipientAddress ?? ''}`,
+        url: undefined,
+        type: '2aee6e5f-091d-444e-96cd-868ba2ddd0e7',
+    }
+
+    const notificationPayload = {
+        notification,
+        accounts,
+    }
     console.log({ notificationPayload })
+
     await fetch('/api/notify', {
         method: 'POST',
         headers: {
